Undo clamping
Clamping was invalid: For ff7 opening credits, the background is 8bit, but expectation is that 255 columns can be read, from adjacent texwindows. Instead, I've allowed the rounding error to happen, and "fixed" the most common case, a coord of -1, to be represented as a signed integer, and then the pixel will just be read from an adjacent texwindow. This still allows a segfault to happen in the case when there is no other texwindow there because it's the boundary of vram.